Big changes to the structure of the data returned by the COYU core functions.
We dispense with mean_sd_data which was originally intended to support plotting but was being used for output.

Instead we supply overall data in the COYUs9Results object and a set of yearly data records in the COYUs9SingleYearResults objects found in each
COYUs9Result$yearly_results member variable

This is much more flexible and reduces duplication of data.

Large changes were required to COYU_results_as_dataframe as a consequence.

In addition plotting was modified to include the function COYU_yearly_plot_data for generating plot data (formerly in the char_results$mean_sd_data
variable) as required.

At the time of this commit COYU_print_results is NON FUNCTIONAL as it still relies on mean_sd_data. This will be fixed shortly